+++ drivers/web-worker-driver/src/jsMain/kotlin/app/cash/sqldelight/driver/worker/expected/JsWorkerResultWithRowCount.kt
@@ -8,7 +8,7 @@ internal class JsWorkerResultWithRowCount(
   private val data: JsWorkerResponse,
 ) : WorkerResultWithRowCount {
   override val rowCount: Long by lazy {
-    when {
+    data.rowCount?.toLong() ?: when {
       data.results.values.isEmpty() -> 0L
       else -> data.results.values[0][0].unsafeCast<Double>().toLong()
     }