+++ drivers/web-worker-driver/src/wasmJsMain/kotlin/app/cash/sqldelight/driver/worker/api/WasmWorkerResponse.kt
@@ -16,6 +16,11 @@ internal external interface WasmWorkerResponse : JsAny {
   var id: Int
 
   /**
+   * Number of rows changed by the request, when supplied by the worker.
+   */
+  var rowCount: JsNumber?
+
+  /**
    * A [WorkerResult] containing any values that were returned by the worker.
    * @see WorkerResult
    */